fix remaining relevant warnings from -Wpedantic

+++ Framework/storage.c
@@ -54,7 +54,7 @@ CnCData read_CNC(char fileName[]) {
   data.columnCount = csv_data.header.columnCount;
 
   size_t num_column_labels = 0;
-  for (int i = 0; i < 100; i++) {
+  for (size_t i = 0; i < 100; i++) {
     if (csv_data.column_labels[i] != NULL) {
       num_column_labels += 1;
     } else {
@@ -72,7 +72,7 @@ CnCData read_CNC(char fileName[]) {
 
   data.columnNames = malloc(data.columnCount * sizeof(char[256]));
 
-  for (int i = 0; i < data.columnCount; i++) {
+  for (size_t i = 0; i < data.columnCount; i++) {
     strcpy(data.columnNames[i], csv_data.column_labels[i]);
   }
 
@@ -82,8 +82,8 @@ CnCData read_CNC(char fileName[]) {
     return data;
   }
 
-  for (int i = 0; i < csv_data.record_count; i++) {
-    for (int j = 0; j < csv_data.field_count[i]; j++) {
+  for (size_t i = 0; i < csv_data.record_count; i++) {
+    for (size_t j = 0; j < csv_data.field_count[i]; j++) {
       data.resultList[(i * csv_data.record_count) + j] = csv_data.records[i][j];
     }
   }
@@ -119,7 +119,7 @@ int write_CNC(char testName[], double resultList[], uint32_t resultCount,
   fprintf(file, "%s\n", testName);
 
   // Writes the column names to the second row of the csv
-  for (int i = 0; i < columnCount; i++) {
+  for (size_t i = 0; i < columnCount; i++) {
     fprintf(file, "%s", columnNames[i]);
     if (i + 1 != columnCount)
       fprintf(file, ",");
@@ -129,7 +129,7 @@ int write_CNC(char testName[], double resultList[], uint32_t resultCount,
   // Writes the resultList in csv format with columnCount enforcing the
   // frequency of line breaks
   uint16_t columnCursor = 0;
-  for (int i = 0; i < resultCount; i++) {
+  for (size_t i = 0; i < resultCount; i++) {
     fprintf(file, "%lf", resultList[i]);
     if (columnCursor < columnCount) {
       columnCursor++;
